(refactor) rename boolean helpers module to extra
