+++ src/registry.rs
@@ -138,15 +138,15 @@ impl<'reg> Registry<'reg> {
         self.register_helper("raw", Box::new(helpers::RAW_HELPER));
         self.register_helper("log", Box::new(helpers::LOG_HELPER));
 
-        self.register_helper("eq", Box::new(helpers::helper_boolean::eq));
-        self.register_helper("ne", Box::new(helpers::helper_boolean::ne));
-        self.register_helper("gt", Box::new(helpers::helper_boolean::gt));
-        self.register_helper("gte", Box::new(helpers::helper_boolean::gte));
-        self.register_helper("lt", Box::new(helpers::helper_boolean::lt));
-        self.register_helper("lte", Box::new(helpers::helper_boolean::lte));
-        self.register_helper("and", Box::new(helpers::helper_boolean::and));
-        self.register_helper("or", Box::new(helpers::helper_boolean::or));
-        self.register_helper("not", Box::new(helpers::helper_boolean::not));
+        self.register_helper("eq", Box::new(helpers::helper_extras::eq));
+        self.register_helper("ne", Box::new(helpers::helper_extras::ne));
+        self.register_helper("gt", Box::new(helpers::helper_extras::gt));
+        self.register_helper("gte", Box::new(helpers::helper_extras::gte));
+        self.register_helper("lt", Box::new(helpers::helper_extras::lt));
+        self.register_helper("lte", Box::new(helpers::helper_extras::lte));
+        self.register_helper("and", Box::new(helpers::helper_extras::and));
+        self.register_helper("or", Box::new(helpers::helper_extras::or));
+        self.register_helper("not", Box::new(helpers::helper_extras::not));
 
         self.register_decorator("inline", Box::new(decorators::INLINE_DECORATOR));
         self
